backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.7.0
xgboost>=2.1.0
tensorflow>=2.20.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
python-dotenv>=1.0.1
slowapi>=0.1.9
pydantic>=2.5.3
requests>=2.31.0
python-multipart>=0.0.9
joblib>=1.4.0
httpx>=0.27.0
